moved exceptions into a package
