+++ api/src/main/java/com/capitalone/dashboard/auth/exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.capitalone.dashboard.service;
+package com.capitalone.dashboard.auth.exceptions;
 
 import com.capitalone.dashboard.model.AuthType;
